refactor: update import paths to use utils from supabase
