+++ client/src/components/base-node.tsx
@@ -1,6 +1,6 @@
 import { forwardRef, HTMLAttributes } from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/utils/supabase/utils";
 
 export const BaseNode = forwardRef<
   HTMLDivElement,
