+++ client/src/components/ui/alert-dialog.tsx
@@ -3,7 +3,7 @@
 import * as React from "react"
 import * as AlertDialogPrimitive from "@radix-ui/react-alert-dialog"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/supabase/utils"
 import { buttonVariants } from "@/components/ui/button"
 
 function AlertDialog({
