+++ client/src/components/ui/popover.tsx
@@ -3,7 +3,7 @@
 import * as React from "react"
 import * as PopoverPrimitive from "@radix-ui/react-popover"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/supabase/utils"
 
 function Popover({
   ...props
